Migrate Compose screen to controller
