+++ presentation/src/main/java/com/moez/QKSMS/feature/compose/ComposeActivity.kt
@@ -18,387 +18,104 @@
  */
 package com.moez.QKSMS.feature.compose
 
-import android.Manifest
-import android.animation.LayoutTransition
-import android.app.Activity
-import android.app.DatePickerDialog
-import android.app.TimePickerDialog
-import android.content.ContentValues
 import android.content.Intent
 import android.net.Uri
-import android.os.Build
 import android.os.Bundle
-import android.provider.ContactsContract
-import android.provider.MediaStore
-import android.text.format.DateFormat
-import android.view.Menu
-import android.view.MenuItem
-import androidx.appcompat.app.AlertDialog
-import androidx.core.app.ActivityCompat
 import androidx.core.view.isVisible
-import androidx.lifecycle.ViewModelProvider
-import androidx.lifecycle.ViewModelProviders
-import com.google.android.flexbox.FlexboxLayoutManager
-import com.google.android.material.snackbar.Snackbar
-import com.jakewharton.rxbinding2.view.clicks
-import com.jakewharton.rxbinding2.widget.textChanges
-import com.moez.QKSMS.R
-import com.moez.QKSMS.common.Navigator
+import com.bluelinelabs.conductor.Conductor
+import com.bluelinelabs.conductor.Router
+import com.bluelinelabs.conductor.RouterTransaction
+import com.google.android.mms.ContentType
 import com.moez.QKSMS.common.base.QkThemedActivity
-import com.moez.QKSMS.common.util.DateFormatter
-import com.moez.QKSMS.common.util.extensions.autoScrollToStart
-import com.moez.QKSMS.common.util.extensions.dismissKeyboard
-import com.moez.QKSMS.common.util.extensions.hideKeyboard
-import com.moez.QKSMS.common.util.extensions.resolveThemeColor
-import com.moez.QKSMS.common.util.extensions.scrapViews
-import com.moez.QKSMS.common.util.extensions.setBackgroundTint
-import com.moez.QKSMS.common.util.extensions.setTint
-import com.moez.QKSMS.common.util.extensions.setVisible
-import com.moez.QKSMS.common.util.extensions.showKeyboard
 import com.moez.QKSMS.common.util.extensions.viewBinding
-import com.moez.QKSMS.databinding.ComposeActivityBinding
-import com.moez.QKSMS.feature.compose.editing.ChipsAdapter
-import com.moez.QKSMS.feature.contacts.ContactsActivity
+import com.moez.QKSMS.databinding.ContainerActivityBinding
 import com.moez.QKSMS.model.Attachment
-import com.moez.QKSMS.model.Recipient
-import com.uber.autodispose.android.lifecycle.scope
-import com.uber.autodispose.autoDisposable
 import dagger.android.AndroidInjection
-import io.reactivex.Observable
-import io.reactivex.subjects.PublishSubject
-import io.reactivex.subjects.Subject
-import java.text.SimpleDateFormat
-import java.util.*
-import javax.inject.Inject
-import kotlin.collections.HashMap
+import java.net.URLDecoder
+import java.nio.charset.Charset
 
-class ComposeActivity : QkThemedActivity(), ComposeView {
+class ComposeActivity : QkThemedActivity() {
 
-    companion object {
-        private const val SelectContactRequestCode = 0
-        private const val TakePhotoRequestCode = 1
-        private const val AttachPhotoRequestCode = 2
-        private const val AttachContactRequestCode = 3
-
-        private const val CameraDestinationKey = "camera_destination"
-    }
-
-    @Inject lateinit var attachmentAdapter: AttachmentAdapter
-    @Inject lateinit var chipsAdapter: ChipsAdapter
-    @Inject lateinit var dateFormatter: DateFormatter
-    @Inject lateinit var messageAdapter: MessagesAdapter
-    @Inject lateinit var navigator: Navigator
-    @Inject lateinit var viewModelFactory: ViewModelProvider.Factory
-
-    override val activityVisibleIntent: Subject<Boolean> = PublishSubject.create()
-    override val chipsSelectedIntent: Subject<HashMap<String, String?>> = PublishSubject.create()
-    override val chipDeletedIntent: Subject<Recipient> by lazy { chipsAdapter.chipDeleted }
-    override val menuReadyIntent: Observable<Unit> = menu.map { Unit }
-    override val optionsItemIntent: Subject<Int> = PublishSubject.create()
-    override val sendAsGroupIntent by lazy { binding.sendAsGroupBackground.clicks() }
-    override val messageClickIntent: Subject<Long> by lazy { messageAdapter.clicks }
-    override val messagePartClickIntent: Subject<Long> by lazy { messageAdapter.partClicks }
-    override val messagesSelectedIntent by lazy { messageAdapter.selectionChanges }
-    override val cancelSendingIntent: Subject<Long> by lazy { messageAdapter.cancelSending }
-    override val attachmentDeletedIntent: Subject<Attachment> by lazy { attachmentAdapter.attachmentDeleted }
-    override val textChangedIntent by lazy { binding.message.textChanges() }
-    override val attachIntent by lazy { Observable.merge(binding.attach.clicks(), binding.attachingBackground.clicks()) }
-    override val cameraIntent by lazy { Observable.merge(binding.camera.clicks(), binding.cameraLabel.clicks()) }
-    override val galleryIntent by lazy { Observable.merge(binding.gallery.clicks(), binding.galleryLabel.clicks()) }
-    override val scheduleIntent by lazy { Observable.merge(binding.schedule.clicks(), binding.scheduleLabel.clicks()) }
-    override val attachContactIntent by lazy { Observable.merge(binding.contact.clicks(), binding.contactLabel.clicks()) }
-    override val attachmentSelectedIntent: Subject<Uri> = PublishSubject.create()
-    override val contactSelectedIntent: Subject<Uri> = PublishSubject.create()
-    override val inputContentIntent by lazy { binding.message.inputContentSelected }
-    override val scheduleSelectedIntent: Subject<Long> = PublishSubject.create()
-    override val changeSimIntent by lazy { binding.sim.clicks() }
-    override val scheduleCancelIntent by lazy { binding.scheduledCancel.clicks() }
-    override val sendIntent by lazy { binding.send.clicks() }
-    override val viewQksmsPlusIntent: Subject<Unit> = PublishSubject.create()
-    override val backPressedIntent: Subject<Unit> = PublishSubject.create()
-
-    private val binding by viewBinding(ComposeActivityBinding::inflate)
-    private val viewModel by lazy { ViewModelProviders.of(this, viewModelFactory)[ComposeViewModel::class.java] }
-
-    private var cameraDestination: Uri? = null
+    private val binding by viewBinding(ContainerActivityBinding::inflate)
+    private lateinit var router: Router
 
     override fun onCreate(savedInstanceState: Bundle?) {
         AndroidInjection.inject(this)
         super.onCreate(savedInstanceState)
         setContentView(binding.root)
-        showBackButton(true)
-        viewModel.bindView(this)
+        toolbar.isVisible = false
 
-        binding.contentView.layoutTransition = LayoutTransition().apply {
-            disableTransitionType(LayoutTransition.CHANGING)
-        }
-
-        chipsAdapter.view = binding.chips
-
-        binding.chips.itemAnimator = null
-        binding.chips.layoutManager = FlexboxLayoutManager(this)
-
-        messageAdapter.autoScrollToStart(binding.messageList)
-        messageAdapter.emptyView = binding.messagesEmpty
-
-        binding.messageList.setHasFixedSize(true)
-        binding.messageList.adapter = messageAdapter
-
-        binding.attachments.adapter = attachmentAdapter
-
-        binding.message.supportsInputContent = true
-
-        theme
-                .doOnNext { binding.loading.setTint(it.theme) }
-                .doOnNext { binding.attach.setBackgroundTint(it.theme) }
-                .doOnNext { binding.attach.setTint(it.textPrimary) }
-                .doOnNext { messageAdapter.theme = it }
-                .autoDisposable(scope())
-                .subscribe()
-
-        window.callback = ComposeWindowCallback(window.callback, this)
-
-        // These theme attributes don't apply themselves on API 21
-        if (Build.VERSION.SDK_INT <= 22) {
-            binding.messageBackground.setBackgroundTint(resolveThemeColor(R.attr.bubbleColor))
+        router = Conductor.attachRouter(this, binding.container, savedInstanceState)
+        if (!router.hasRootController()) {
+            router.setRoot(RouterTransaction.with(ComposeController(
+                    getQuery(), getThreadId(), getAddresses(), getSharedText(), getSharedAttachments())))
         }
     }
 
-    override fun onStart() {
-        super.onStart()
-        activityVisibleIntent.onNext(true)
-    }
-
-    override fun onPause() {
-        super.onPause()
-        activityVisibleIntent.onNext(false)
-    }
-
-    override fun render(state: ComposeState) {
-        if (state.hasError) {
-            finish()
-            return
-        }
-
-        threadId.onNext(state.threadId)
-
-        title = when {
-            state.selectedMessages > 0 -> getString(R.string.compose_title_selected, state.selectedMessages)
-            state.query.isNotEmpty() -> state.query
-            else -> state.conversationtitle
-        }
-
-        binding.toolbarSubtitle.setVisible(state.query.isNotEmpty())
-        binding.toolbarSubtitle.text = getString(R.string.compose_subtitle_results, state.searchSelectionPosition,
-                state.searchResults)
-
-        binding.toolbarTitle.setVisible(!state.editingMode)
-        binding.chips.setVisible(state.editingMode)
-        binding.composeBar.setVisible(!state.loading)
-
-        // Don't set the adapters unless needed
-        if (state.editingMode && binding.chips.adapter == null) binding.chips.adapter = chipsAdapter
-
-        binding.toolbar.menu.findItem(R.id.add)?.isVisible = state.editingMode
-        binding.toolbar.menu.findItem(R.id.call)?.isVisible = !state.editingMode && state.selectedMessages == 0
-                && state.query.isEmpty()
-        binding.toolbar.menu.findItem(R.id.info)?.isVisible = !state.editingMode && state.selectedMessages == 0
-                && state.query.isEmpty()
-        binding.toolbar.menu.findItem(R.id.copy)?.isVisible = !state.editingMode && state.selectedMessages > 0
-        binding.toolbar.menu.findItem(R.id.details)?.isVisible = !state.editingMode && state.selectedMessages == 1
-        binding.toolbar.menu.findItem(R.id.delete)?.isVisible = !state.editingMode && state.selectedMessages > 0
-        binding.toolbar.menu.findItem(R.id.forward)?.isVisible = !state.editingMode && state.selectedMessages == 1
-        binding.toolbar.menu.findItem(R.id.previous)?.isVisible = state.selectedMessages == 0 && state.query.isNotEmpty()
-        binding.toolbar.menu.findItem(R.id.next)?.isVisible = state.selectedMessages == 0 && state.query.isNotEmpty()
-        binding.toolbar.menu.findItem(R.id.clear)?.isVisible = state.selectedMessages == 0 && state.query.isNotEmpty()
-
-        chipsAdapter.data = state.selectedChips
-
-        binding.loading.setVisible(state.loading)
-
-        binding.sendAsGroup.setVisible(state.editingMode && state.selectedChips.size >= 2)
-        binding.sendAsGroupSwitch.isChecked = state.sendAsGroup
-
-        binding.messageList.setVisible(!state.editingMode || state.sendAsGroup || state.selectedChips.size == 1)
-        messageAdapter.data = state.messages
-        messageAdapter.highlight = state.searchSelectionId
-
-        binding.scheduledGroup.isVisible = state.scheduled != 0L
-        binding.scheduledTime.text = dateFormatter.getScheduledTimestamp(state.scheduled)
-
-        binding.attachments.setVisible(state.attachments.isNotEmpty())
-        attachmentAdapter.data = state.attachments
-
-        binding.attach.animate().rotation(if (state.attaching) 135f else 0f).start()
-        binding.attaching.isVisible = state.attaching
-
-        binding.counter.text = state.remaining
-        binding.counter.setVisible(binding.counter.text.isNotBlank())
-
-        binding.sim.setVisible(state.subscription != null)
-        binding.sim.contentDescription = getString(R.string.compose_sim_cd, state.subscription?.displayName)
-        binding.simIndex.text = state.subscription?.simSlotIndex?.plus(1)?.toString()
-
-        binding.send.isEnabled = state.canSend
-        binding.send.imageAlpha = if (state.canSend) 255 else 128
-    }
-
-    override fun clearSelection() = messageAdapter.clearSelection()
-
-    override fun showDetails(details: String) {
-        AlertDialog.Builder(this)
-                .setTitle(R.string.compose_details_title)
-                .setMessage(details)
-                .setCancelable(true)
-                .show()
-    }
-
-    override fun requestDefaultSms() {
-        navigator.showDefaultSmsDialog(this)
-    }
-
-    override fun requestStoragePermission() {
-        ActivityCompat.requestPermissions(this, arrayOf(Manifest.permission.WRITE_EXTERNAL_STORAGE), 0)
-    }
-
-    override fun requestSmsPermission() {
-        ActivityCompat.requestPermissions(this, arrayOf(
-                Manifest.permission.READ_SMS,
-                Manifest.permission.SEND_SMS), 0)
-    }
-
-    override fun requestDatePicker() {
-        val calendar = Calendar.getInstance()
-        DatePickerDialog(this, DatePickerDialog.OnDateSetListener { _, year, month, day ->
-            TimePickerDialog(this, TimePickerDialog.OnTimeSetListener { _, hour, minute ->
-                calendar.set(Calendar.YEAR, year)
-                calendar.set(Calendar.MONTH, month)
-                calendar.set(Calendar.DAY_OF_MONTH, day)
-                calendar.set(Calendar.HOUR_OF_DAY, hour)
-                calendar.set(Calendar.MINUTE, minute)
-                scheduleSelectedIntent.onNext(calendar.timeInMillis)
-            }, calendar.get(Calendar.HOUR_OF_DAY), calendar.get(Calendar.MINUTE), DateFormat.is24HourFormat(this))
-                    .show()
-        }, calendar.get(Calendar.YEAR), calendar.get(Calendar.MONTH), calendar.get(Calendar.DAY_OF_MONTH)).show()
-
-        // On some devices, the keyboard can cover the date picker
-        binding.message.hideKeyboard()
-    }
-
-    override fun requestContact() {
-        val intent = Intent(Intent.ACTION_PICK)
-                .setType(ContactsContract.CommonDataKinds.Phone.CONTENT_TYPE)
-
-        startActivityForResult(Intent.createChooser(intent, null), AttachContactRequestCode)
-    }
-
-    override fun showContacts(sharing: Boolean, chips: List<Recipient>) {
-        binding.message.hideKeyboard()
-        val serialized = HashMap(chips.associate { chip -> chip.address to chip.contact?.lookupKey })
-        val intent = Intent(this, ContactsActivity::class.java)
-                .putExtra(ContactsActivity.SharingKey, sharing)
-                .putExtra(ContactsActivity.ChipsKey, serialized)
-        startActivityForResult(intent, SelectContactRequestCode)
-    }
-
-    override fun themeChanged() {
-        binding.messageList.scrapViews()
-    }
-
-    override fun showKeyboard() {
-        binding.message.postDelayed({
-            binding.message.showKeyboard()
-        }, 200)
-    }
-
-    override fun requestCamera() {
-        cameraDestination = SimpleDateFormat("yyyyMMdd_HHmmss", Locale.getDefault()).format(Date())
-                .let { timestamp -> ContentValues().apply { put(MediaStore.Images.Media.TITLE, timestamp) } }
-                .let { cv -> contentResolver.insert(MediaStore.Images.Media.EXTERNAL_CONTENT_URI, cv) }
-
-        val intent = Intent(MediaStore.ACTION_IMAGE_CAPTURE)
-                .putExtra(MediaStore.EXTRA_OUTPUT, cameraDestination)
-        startActivityForResult(Intent.createChooser(intent, null), TakePhotoRequestCode)
-    }
-
-    override fun requestGallery() {
-        val intent = Intent(Intent.ACTION_PICK)
-                .putExtra(Intent.EXTRA_ALLOW_MULTIPLE, true)
-                .addFlags(Intent.FLAG_GRANT_PERSISTABLE_URI_PERMISSION)
-                .putExtra(Intent.EXTRA_LOCAL_ONLY, false)
-                .addFlags(Intent.FLAG_GRANT_READ_URI_PERMISSION)
-                .setType("image/*")
-        startActivityForResult(Intent.createChooser(intent, null), AttachPhotoRequestCode)
-    }
-
-    override fun setDraft(draft: String) {
-        binding.message.setText(draft)
-        binding.message.setSelection(draft.length)
-    }
-
-    override fun scrollToMessage(id: Long) {
-        messageAdapter.data?.second
-                ?.indexOfLast { message -> message.id == id }
-                ?.takeIf { position -> position != -1 }
-                ?.let(binding.messageList::scrollToPosition)
-    }
-
-    override fun showQksmsPlusSnackbar(message: Int) {
-        Snackbar.make(binding.contentView, message, Snackbar.LENGTH_LONG).run {
-            setAction(R.string.button_more) { viewQksmsPlusIntent.onNext(Unit) }
-            setActionTextColor(colors.theme().theme)
-            show()
+    override fun onBackPressed() {
+        if (!router.handleBack()) {
+            super.onBackPressed()
         }
     }
 
-    override fun onCreateOptionsMenu(menu: Menu?): Boolean {
-        menuInflater.inflate(R.menu.compose, menu)
-        return super.onCreateOptionsMenu(menu)
+    private fun getQuery(): String = intent.extras?.getString("query") ?: ""
+
+    private fun getThreadId(): Long = intent.extras?.getLong("threadId") ?: 0L
+
+    private fun getAddresses(): List<String> {
+        return intent
+                ?.decodedDataString()
+                ?.substringAfter(':') // Remove scheme
+                ?.replaceAfter("?", "") // Remove query
+                ?.split(",", ";")
+                ?: listOf()
     }
 
-    override fun onOptionsItemSelected(item: MenuItem): Boolean {
-        optionsItemIntent.onNext(item.itemId)
-        return true
+    private fun getSharedText(): String {
+        var subject = intent.getStringExtra(Intent.EXTRA_SUBJECT) ?: "";
+        if (subject != "") {
+            subject += "\n"
+        }
+
+        return subject + (intent.extras?.getString(Intent.EXTRA_TEXT)
+                ?: intent.extras?.getString("sms_body")
+                ?: intent?.decodedDataString()
+                        ?.substringAfter('?') // Query string
+                        ?.split(',')
+                        ?.firstOrNull { param -> param.startsWith("body") }
+                        ?.substringAfter('=')
+                ?: "")
     }
 
-    override fun getColoredMenuItems(): List<Int> {
-        return super.getColoredMenuItems() + R.id.call
-    }
+    private fun getSharedAttachments(): List<Attachment> {
+        val uris = mutableListOf<Uri>()
+        intent.getParcelableExtra<Uri>(Intent.EXTRA_STREAM)?.run(uris::add)
+        intent.getParcelableArrayListExtra<Uri>(Intent.EXTRA_STREAM)?.run(uris::addAll)
+        return uris.mapNotNull { uri ->
+            val mimeType = contentResolver.getType(uri)
+            when {
+                ContentType.isImageType(mimeType) -> {
+                    Attachment.Image(uri)
+                }
 
-    override fun onActivityResult(requestCode: Int, resultCode: Int, data: Intent?) {
-        when {
-            requestCode == SelectContactRequestCode -> {
-                chipsSelectedIntent.onNext(data?.getSerializableExtra(ContactsActivity.ChipsKey)
-                        ?.let { serializable -> serializable as? HashMap<String, String?> }
-                        ?: hashMapOf())
+                ContentType.TEXT_VCARD.equals(mimeType, true) -> {
+                    val inputStream = contentResolver.openInputStream(uri)
+                    val text = inputStream?.reader(Charset.forName("utf-8"))?.readText()
+                    text?.let(Attachment::Contact)
+                }
+
+                else -> null
             }
-            requestCode == TakePhotoRequestCode && resultCode == Activity.RESULT_OK -> {
-                cameraDestination?.let(attachmentSelectedIntent::onNext)
-            }
-            requestCode == AttachPhotoRequestCode && resultCode == Activity.RESULT_OK -> {
-                data?.clipData?.itemCount
-                        ?.let { count -> 0 until count }
-                        ?.mapNotNull { i -> data.clipData?.getItemAt(i)?.uri }
-                        ?.forEach(attachmentSelectedIntent::onNext)
-                        ?: data?.data?.let(attachmentSelectedIntent::onNext)
-            }
-            requestCode == AttachContactRequestCode && resultCode == Activity.RESULT_OK -> {
-                data?.data?.let(contactSelectedIntent::onNext)
-            }
-            else -> super.onActivityResult(requestCode, resultCode, data)
         }
     }
 
-    override fun onSaveInstanceState(outState: Bundle) {
-        outState.putParcelable(CameraDestinationKey, cameraDestination)
-        super.onSaveInstanceState(outState)
+    // The dialer app on Oreo sends a URL encoded string, make sure to decode it
+    private fun Intent.decodedDataString(): String? {
+        val data = data?.toString()
+        if (data?.contains('%') == true) {
+            return URLDecoder.decode(data, "UTF-8")
+        }
+        return data
     }
-
-    override fun onRestoreInstanceState(savedInstanceState: Bundle?) {
-        cameraDestination = savedInstanceState?.getParcelable(CameraDestinationKey)
-        super.onRestoreInstanceState(savedInstanceState)
-    }
-
-    override fun onBackPressed() = backPressedIntent.onNext(Unit)
 
 }
