+++ presentation/src/main/java/com/moez/QKSMS/feature/compose/ComposePresenter.kt
@@ -26,7 +26,7 @@ import android.telephony.SmsMessage
 import androidx.core.content.getSystemService
 import com.moez.QKSMS.R
 import com.moez.QKSMS.common.Navigator
-import com.moez.QKSMS.common.base.QkViewModel
+import com.moez.QKSMS.common.base.QkPresenter
 import com.moez.QKSMS.common.util.BillingManager
 import com.moez.QKSMS.common.util.ClipboardUtils
 import com.moez.QKSMS.common.util.MessageDetailsFormatter
@@ -73,7 +73,7 @@ import java.util.*
 import javax.inject.Inject
 import javax.inject.Named
 
-class ComposeViewModel @Inject constructor(
+class ComposePresenter @Inject constructor(
     @Named("query") private val query: String,
     @Named("threadId") private val threadId: Long,
     @Named("addresses") private val addresses: List<String>,
@@ -97,7 +97,7 @@ class ComposeViewModel @Inject constructor(
     private val retrySending: RetrySending,
     private val sendMessage: SendMessage,
     private val subscriptionManager: SubscriptionManagerCompat
-) : QkViewModel<ComposeView, ComposeState>(ComposeState(
+) : QkPresenter<ComposeView, ComposeState>(ComposeState(
         editingMode = threadId == 0L && addresses.isEmpty(),
         threadId = threadId,
         query = query)
@@ -227,8 +227,8 @@ class ComposeViewModel @Inject constructor(
         }.subscribe()
     }
 
-    override fun bindView(view: ComposeView) {
-        super.bindView(view)
+    override fun bindIntents(view: ComposeView) {
+        super.bindIntents(view)
 
         val sharing = sharedText.isNotEmpty() || sharedAttachments.isNotEmpty()
         if (shouldShowContacts) {
