+++ presentation/src/main/java/com/moez/QKSMS/feature/compose/ComposeView.kt
@@ -22,12 +22,13 @@ import android.net.Uri
 import androidx.annotation.StringRes
 import androidx.core.view.inputmethod.InputContentInfoCompat
 import com.moez.QKSMS.common.base.QkView
+import com.moez.QKSMS.common.base.QkViewContract
 import com.moez.QKSMS.model.Attachment
 import com.moez.QKSMS.model.Recipient
 import io.reactivex.Observable
 import io.reactivex.subjects.Subject
 
-interface ComposeView : QkView<ComposeState> {
+interface ComposeView : QkViewContract<ComposeState> {
 
     val activityVisibleIntent: Observable<Boolean>
     val chipsSelectedIntent: Subject<HashMap<String, String?>>
