+++ presentation/src/main/java/com/moez/QKSMS/injection/AppComponent.kt
@@ -34,6 +34,7 @@ import com.moez.QKSMS.feature.blocking.manager.BlockingManagerController
 import com.moez.QKSMS.feature.blocking.messages.BlockedMessagesController
 import com.moez.QKSMS.feature.blocking.numbers.BlockedNumbersController
 import com.moez.QKSMS.feature.compose.editing.DetailedChipView
+import com.moez.QKSMS.feature.compose.injection.ComposeComponent
 import com.moez.QKSMS.feature.contacts.injection.ContactsComponent
 import com.moez.QKSMS.feature.conversationinfo.injection.ConversationInfoComponent
 import com.moez.QKSMS.feature.gallery.injection.GalleryComponent
@@ -61,6 +62,7 @@ import javax.inject.Singleton
     ServiceBuilderModule::class])
 interface AppComponent {
 
+    fun composeBuilder(): ComposeComponent.Builder
     fun contactsBuilder(): ContactsComponent.Builder
     fun conversationInfoBuilder(): ConversationInfoComponent.Builder
     fun galleryBuilder(): GalleryComponent.Builder
